% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_manifest}
\alias{create_manifest}
\title{Create a minimal TOML manifest file}
\usage{
create_manifest(
  path,
  name = "Project",
  version = manifest_version(),
  r_version = "*",
  ...
)
}
\arguments{
\item{path}{File path to write to. If missing, a temporary file will be used.}

\item{name}{Optional project name. Defaults to \code{Project}.}

\item{version}{Optional project version. Defaults to \code{manifest_version()}.}

\item{r_version}{Optional R version settings. Defaults to \code{'*'}.}
}
\value{
Invisibly returns the written path.
}
\description{
This creates a minimal TOML file with a \verb{[manifesto]}, \verb{[project]},
\verb{[environment]}, and empty \verb{[dependencies]} section.
}
\examples{
path <- create_manifest(
  dependencies = list(dplyr = '>= 1.0.0'),
  'suggests-dependencies' = list(testthat = '>= 3.0.0')
)
}
