% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version.R
\name{current_r_version}
\alias{current_r_version}
\title{Get the current R version}
\usage{
current_r_version()
}
\value{
A character string representing the full R version.
}
\description{
The version is returned as a string in the format \code{"major.minor.patch"} (e.g., \code{"4.4.0"}),
using components from \code{R.version}.
}
\examples{
current_r_version()
}
