% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{manifest_all_groups}
\alias{manifest_all_groups}
\title{Return all defined optional dependency groups in a manifest file}
\usage{
manifest_all_groups(path = "rproject.toml")
}
\arguments{
\item{path}{Path to the \code{rproject.toml} file.}
}
\value{
A character vector of group names (without "-dependencies")
}
\description{
Return all defined optional dependency groups in a manifest file
}
\examples{
manifest_all_groups(path = system.file(package = 'manifesto', 'minimal.toml'))
}
