% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{manifest_peek}
\alias{manifest_peek}
\title{Peek into a manifest file}
\usage{
manifest_peek(path = "rproject.toml")
}
\arguments{
\item{path}{Path to the \code{rproject.toml} file. Defaults to "rproject.toml" in the current directory.}
}
\value{
Invisibly returns a list with parsed contents.
}
\description{
Peek into a manifest file
}
\examples{
manifest_peek(system.file('complex.toml', package = 'manifesto'))
}
