% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{manifest_version}
\alias{manifest_version}
\title{Return the current version of the manifesto package}
\usage{
manifest_version()
}
\value{
A character version string.
}
\description{
Return the current version of the manifesto package
}
\examples{
manifest_version()
}
