% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_manifest}
\alias{validate_manifest}
\title{Validate a manifesto manifest file}
\usage{
validate_manifest(path = "rproject.toml", groups = NULL)
}
\arguments{
\item{path}{Path to the \code{rproject.toml} file.}

\item{groups}{Optional character vector of dependency groups to include.}
}
\value{
Invisibly returns \code{TRUE} if the manifest is valid; otherwise, stops with an error.
}
\description{
Validate a manifesto manifest file
}
\examples{
validate_manifest(path = system.file(package = 'manifesto', 'minimal.toml'))
}
