% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{manifest_check_system}
\alias{manifest_check_system}
\title{Check for system dependencies}
\usage{
manifest_check_system(path = "rproject.toml")
}
\arguments{
\item{path}{Path to the \code{rproject.toml} file.}
}
\value{
A \code{data.frame} reporting the system dependency, and its status.
}
\description{
This function checks for the presence of system dependencies listed in the
\verb{[system-dependencies]} section of the manifest file.
}
\examples{
path <- manifest_create(`system-dependencies` = list(git = '*'))
manifest_check_system(path)
}
