% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{manifest_create}
\alias{manifest_create}
\title{Create a minimal TOML manifest file}
\usage{
manifest_create(
  path,
  project_name = "Project",
  project_version = "0.0.1",
  manifesto_version = manifest_version(),
  r_version = "*",
  ...
)
}
\arguments{
\item{path}{File path to write to. If missing, a temporary file will be used.}

\item{project_name}{Optional project name. Defaults to \code{Project}.}

\item{project_version}{Optional project version. Defaults to \verb{0.0.1}.}

\item{r_version}{Optional R version settings. Defaults to \code{'*'}.}

\item{...}{Additional named arguments to add to the manifest. These will be
added to the top-level of the TOML file.}

\item{manifest_version}{Optional \code{manifesto} version. Defaults to the current
package version.}
}
\value{
Invisibly returns the written path.
}
\description{
This creates a minimal TOML file with a \verb{[manifesto]}, \verb{[project]},
\verb{[environment]}, and empty \verb{[dependencies]} section.
}
\examples{
path <- manifest_create(
  dependencies = list(dplyr = '>= 1.0.0'),
  'suggests-dependencies' = list(testthat = '>= 3.0.0')
)
}
