% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_manifest}
\alias{parse_manifest}
\title{Parse a manifesto rproject.toml manifest file}
\usage{
parse_manifest(path = "rproject.toml", groups = NULL)
}
\arguments{
\item{path}{Path to the \code{rproject.toml} file.}

\item{groups}{Optional character vector of dependency groups to include.}
}
\value{
A character vector of resolved package references.
}
\description{
By default, \code{groups = NULL}, which will only install core dependencies.
If you want to include all optional groups, set \code{groups = 'all'}.
}
\examples{
parse_manifest(path = system.file(package = 'manifesto', 'minimal.toml'))
}
